chore: Release 8.2.1

+++ lib/GroupBackend.php
@@ -45,6 +45,8 @@ class GroupBackend extends ABackend implements
 	/** @var array<string, string|null> */
 	private array $groupCache = [];
 
+	protected ?UserBackend $userBackend = null;
+
 	public const TABLE_GROUPS = 'user_saml_groups';
 	public const TABLE_MEMBERS = 'user_saml_group_members';
 
@@ -54,8 +56,16 @@ class GroupBackend extends ABackend implements
 	) {
 	}
 
+	private function getUserBackend(): UserBackend {
+		return $this->userBackend ?? Server::get(UserBackend::class);
+	}
+
 	#[\Override]
 	public function inGroup($uid, $gid): bool {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group membership
+			return false;
+		}
 		$qb = $this->dbc->getQueryBuilder();
 		$stmt = $qb->select('gid')
 			->from(self::TABLE_MEMBERS)
@@ -75,6 +85,10 @@ class GroupBackend extends ABackend implements
 	 */
 	#[\Override]
 	public function getUserGroups($uid): array {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for the list of groups
+			return [];
+		}
 		$qb = $this->dbc->getQueryBuilder();
 		$cursor = $qb->select('gu.gid', 'g.displayname')
 			->from(self::TABLE_MEMBERS, 'gu')
@@ -98,6 +112,10 @@ class GroupBackend extends ABackend implements
 	 */
 	#[\Override]
 	public function getGroups(string $search = '', ?int $limit = null, ?int $offset = null): array {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for the list of groups
+			return [];
+		}
 		$query = $this->dbc->getQueryBuilder();
 		$query->select('gid', 'displayname')
 			->from(self::TABLE_GROUPS)
@@ -132,6 +150,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function groupExists($gid): bool {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for the list of groups
+			return false;
+		}
+
 		if (isset($this->groupCache[$gid])) {
 			return true;
 		}
@@ -154,6 +177,10 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function groupsExists(array $gids): array {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for the list of groups
+			return [];
+		}
 		$notFoundGids = [];
 		$existingGroups = [];
 
@@ -186,6 +213,10 @@ class GroupBackend extends ABackend implements
 	}
 
 	public function groupExistsWithDifferentGid(string $samlGid): ?string {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for the list of groups
+			return false;
+		}
 		$qb = $this->dbc->getQueryBuilder();
 		$cursor = $qb->select('gid')
 			->from(self::TABLE_GROUPS)
@@ -209,6 +240,11 @@ class GroupBackend extends ABackend implements
 	 */
 	#[\Override]
 	public function usersInGroup($gid, $search = '', $limit = -1, $offset = 0): array {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group membership
+			return [];
+		}
+
 		$query = $this->dbc->getQueryBuilder();
 		$query->selectDistinct('m.uid')
 			->from(self::TABLE_MEMBERS, 'm')
@@ -245,6 +281,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function createGroup(string $name, ?string $samlGid = null): ?string {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We create groups on the auto-provisioning backend
+			return null;
+		}
+
 		$gid = $this->computeGid($name);
 
 		try {
@@ -277,6 +318,11 @@ class GroupBackend extends ABackend implements
 	 */
 	#[\Override]
 	public function addToGroup(string $uid, string $gid): bool {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We add user to groups on the auto-provisioning backend
+			return false;
+		}
+
 		if ($this->inGroup($uid, $gid)) {
 			return true;
 		}
@@ -291,6 +337,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function removeFromGroup(string $uid, string $gid): bool {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We remove user to groups on the auto-provisioning backend
+			return false;
+		}
+
 		$qb = $this->dbc->getQueryBuilder();
 		$rows = $qb->delete(self::TABLE_MEMBERS)
 			->where($qb->expr()->eq('uid', $qb->createNamedParameter($uid)))
@@ -302,6 +353,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function countUsersInGroup(string $gid, string $search = ''): int {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group membership
+			return 0;
+		}
+
 		$query = $this->dbc->getQueryBuilder();
 		$query->select(
 			$query->createFunction('COUNT(DISTINCT ' . $query->getColumnName('uid', 'm') . ')')
@@ -326,6 +382,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function deleteGroup(string $gid): bool {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group management
+			return 0;
+		}
+
 		$query = $this->dbc->getQueryBuilder();
 
 		try {
@@ -358,6 +419,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function getDisplayName(string $gid): string {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group management
+			return '';
+		}
+
 		if (!isset($this->groupCache[$gid])) {
 			$this->getGroups($gid);
 		}
@@ -367,6 +433,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function setDisplayName(string $gid, string $displayName): bool {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group management
+			return false;
+		}
+
 		if (!$this->groupExists($gid)) {
 			return false;
 		}
@@ -391,6 +462,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function getGroupDetails(string $gid): array {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group management
+			return [];
+		}
+
 		if (!$this->groupExists($gid)) {
 			return [];
 		}
@@ -400,6 +476,11 @@ class GroupBackend extends ABackend implements
 
 	#[\Override]
 	public function getGroupsDetails(array $gids): array {
+		if (!$this->getUserBackend()->autoprovisionAllowed()) {
+			// We rely on the auto-provisioning backend for group management
+			return [];
+		}
+
 		$notFoundGids = [];
 		$details = [];
